+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+- Added Consolas to the diagram font list.
+
 ## [0.3.0-alpha.3] - 2021-11-28
 
 - Fixed the UAX #11 width calculation of texts including characters that are translated to XML entity references by `svgbob`.